fix: control planes config v2 spec (#683)
* fix: add next field to the list core entities response

We already return this field in the response. The schema
doesn't capture this correctly. Adding the next field
in the schema.

* fix: remove ControlPlaneId parameter from generated.yaml

we define controlPlaneId and ControlPlaneId separately in
two different specs and then combine them. This results
in two controlPlaneId parameters in the combined spec.

This results in an error when generating server side code.
Removing this from generated.yaml to be able to generate
server side code.